+++ shorts-api/src/main/kotlin/com/mashup/shorts/domain/newscard/NewsCardController.kt
@@ -39,5 +39,4 @@ class NewsCardController(
         )
     }
 
-
 }
