+++ shorts-crawler/src/main/kotlin/com/mashup/shorts/core/util/CrawlerContentConverter.kt
@@ -1,0 +1,16 @@
+package com.mashup.shorts.core.util
+
+import org.springframework.stereotype.Component
+import com.mashup.shorts.core.const.NewsDOMClassNameConst
+
+@Component
+class CrawlerContentConverter {
+
+    internal fun convertHeadLine(headLineFlag: Boolean): String {
+        return if (headLineFlag) {
+            NewsDOMClassNameConst.HEADLINE
+        } else {
+            NewsDOMClassNameConst.NORMAL
+        }
+    }
+}
