+++ shorts-crawler/src/main/kotlin/com/mashup/shorts/core/util/CrawlerContentFilter.kt
@@ -1,0 +1,19 @@
+package com.mashup.shorts.core.util
+
+import org.springframework.stereotype.Component
+
+@Component
+class CrawlerContentFilter {
+
+    internal fun filterSquareBracket(target: String): String {
+        return target
+            .replace("[", "")
+            .replace("]", "")
+    }
+
+    internal fun filterImageLinkForm(rawImageLink: String): String {
+        return rawImageLink
+            .substringAfter("data-src=\"")
+            .substringBefore("\"")
+    }
+}
